requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
blas=1.0=mkl
brotli=1.0.9=ha925a31_2
brotlipy=0.7.0=py38h294d835_1004
ca-certificates=2022.6.15=h5b45459_0
certifi=2022.6.15=py38haa244fe_0
cffi=1.15.1=py38hd8c33c5_0
charset-normalizer=2.1.0=pyhd8ed1ab_0
cloudpickle=2.2.0=pypi_0
colorama=0.4.5=pypi_0
cryptography=37.0.4=py38hb7941b4_0
cudatoolkit=11.6.0=hc0ea762_10
cycler=0.11.0=pyhd3eb1b0_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.10.4=hd328e21_0
glib=2.69.1=h5dc1a3c_1
gst-plugins-base=1.18.5=h9e645db_0
gstreamer=1.18.5=hd78058f_0
gym=0.26.1=pypi_0
gym-notices=0.0.8=pypi_0
gym-tetris=3.0.4=pypi_0
icu=58.2=ha925a31_3
idna=3.3=pyhd8ed1ab_0
importlib-metadata=4.12.0=pypi_0
intel-openmp=2021.4.0=haa95532_3556
jpeg=9e=h2bbff1b_0
kiwisolver=1.4.2=py38hd77b12b_0
libclang=12.0.0=default_h627e005_2
libffi=3.4.2=hd77b12b_4
libiconv=1.16=h2bbff1b_2
libogg=1.3.5=h2bbff1b_1
libpng=1.6.37=h2a8f88b_0
libtiff=4.2.0=he0120a3_1
libuv=1.44.2=h8ffe710_0
libvorbis=1.3.7=he774522_0
libwebp=1.2.2=h2bbff1b_0
libxml2=2.9.14=h0ad7f3c_0
libxslt=1.1.35=h2bbff1b_0
lz4-c=1.9.3=h2bbff1b_1
matplotlib=3.5.1=py38haa95532_1
matplotlib-base=3.5.1=py38hd77b12b_1
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py38h2bbff1b_0
mkl_fft=1.3.1=py38h277e83a_0
mkl_random=1.2.2=py38hf11a4ad_0
munkres=1.1.4=py_0
nes-py=8.2.1=pypi_0
numpy=1.23.1=py38h7a0a035_0
numpy-base=1.23.1=py38hca35cd5_0
openssl=1.1.1q=h8ffe710_0
packaging=21.3=pyhd3eb1b0_0
pcre=8.45=hd77b12b_0
pillow=9.2.0=py38hdc2b20a_1
pip=22.1.2=py38haa95532_0
ply=3.11=py38_0
pycparser=2.21=pyhd8ed1ab_0
pygame=2.1.3.dev2=pypi_0
pyglet=1.5.21=pypi_0
pyopenssl=22.0.0=pyhd8ed1ab_0
pyparsing=3.0.4=pyhd3eb1b0_0
pyqt=5.15.7=py38hd77b12b_0
pyqt5-sip=12.11.0=py38hd77b12b_0
pysocks=1.7.1=py38haa244fe_5
python=3.8.13=h6244533_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.8=2_cp38
pytorch=1.12.0=py3.8_cuda11.6_cudnn8_0
pytorch-mutex=1.0=cuda
qt-main=5.15.2=he8e5bd7_4
qt-webengine=5.15.9=hb9a9bb5_4
qtwebkit=5.212=h3ad3cdb_4
requests=2.28.1=pyhd8ed1ab_0
setuptools=61.2.0=py38haa95532_0
sip=6.6.2=py38hd77b12b_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.38.5=h2bbff1b_0
tk=8.6.12=h2bbff1b_0
toml=0.10.2=pyhd3eb1b0_0
torchaudio=0.12.0=py38_cu116
torchvision=0.13.0=py38_cu116
tornado=6.1=py38h2bbff1b_0
tqdm=4.64.1=pypi_0
typing_extensions=4.3.0=pyha770c72_0
urllib3=1.26.11=pyhd8ed1ab_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.37.1=pyhd3eb1b0_0
win_inet_pton=1.1.0=py38haa244fe_4
wincertstore=0.2=py38haa95532_2
xz=5.2.5=h8cc25b3_1
zipp=3.8.1=pypi_0
zlib=1.2.12=h8cc25b3_2
zstd=1.5.2=h19a0ad4_0
